+++ core/config.ts
@@ -6,6 +6,12 @@ function toA<T>(list?: T[]): T[] {
 	return list;
 }
 
+export type Options = {
+	dry_run: boolean;
+	serverUrl: URL;
+	serverAddress: string;
+}
+
 export type Engine = { name: string, url: string };
 export type Layout = { name: string, filepath: string, engine: string };
 export type Static = { pattern: string, replace: string };
@@ -13,7 +19,6 @@ export type Dynamic = { pattern: string, replace: string, engine: string, layout
 export type Config = {
 	src?: string;
 	dest?: string;
-	baseurl?: string;
 	index?: string[];
 	datas?: string[];
 	include?: string;
@@ -22,13 +27,10 @@ export type Config = {
 	statics?: Static[];
 	dynamics?: Dynamic[];
 	ignores?: string[];
-	dry_run?: boolean;
-	deploy?: string[];
 }
 export function mergeConfig(config1: Config, config2: Config): Config {
 	const src = config2.src || config1.src;
 	const dest = config2.dest || config1.dest;
-	const baseurl = config2.baseurl || config1.baseurl;
 	const index = toA(config1.index).concat(toA(config2.index));
 	const datas = toA(config1.datas).concat(toA(config2.datas));
 	const include = config2.include || config1.include;
@@ -37,8 +39,6 @@ export function mergeConfig(config1: Config, config2: Config): Config {
 	const statics = toA(config1.statics).concat(toA(config2.statics));
 	const dynamics = toA(config1.dynamics).concat(toA(config2.dynamics));
 	const ignores = toA(config1.ignores).concat(toA(config2.ignores));
-	const dry_run = config2.dry_run || config1.dry_run;
-	const deploy = toA(config1.deploy).concat(toA(config2.deploy));
-	const config: Config = { src, dest, baseurl, index, datas, include, engines, layouts, statics, dynamics, ignores, dry_run, deploy };
+	const config: Config = { src, dest, index, datas, include, engines, layouts, statics, dynamics, ignores };
 	return config;
 }